Battery status check time display in user friendly form, cleanup.

+++ app/src/main/java/pl/appnode/roy/BatteryItem.java
@@ -6,7 +6,7 @@ package pl.appnode.roy;
 public class BatteryItem {
     String batteryDeviceId;
     String batteryDeviceName;
-    float batteryCheckTime;
+    long batteryCheckTime;
     int batteryLevel;
     int batteryPluggedStatus;
     int batteryChargingStatus;
